Move parse_argument to utils.
