+++ src/utils/__init__.py
@@ -1,9 +1,10 @@
 from .config import get_config
 from .cache import get_cache
-from .utils import std_out_err_redirect_tqdm
+from .utils import std_out_err_redirect_tqdm, parse_arguments
 
 __all__ = [
     "get_config",
     "get_cache",
     "std_out_err_redirect_tqdm",
+    "parse_arguments",
 ]
